pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "clinical-rag-pipeline"
version = "1.0.0"
description = "ICU patient risk scoring and clinical LLM RAG pipeline built on MIMIC-III"
requires-python = ">=3.11"

# Base dependencies — used everywhere (API, Airflow workers, tests)
dependencies = [
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.5",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "anthropic>=0.28",
    "openai>=1.30",
    "prometheus-client>=0.20",
    "httpx>=0.27",
]
spark = [
    "pyspark>=3.5.1",
    "delta-spark>=2.4",
]
embeddings = [
    "openai>=1.30",
    "sentence-transformers>=3.0",
    "torch>=2.2; sys_platform != 'win32'",
]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.10",
    "coverage[toml]>=7",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # handled by formatter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["api", "embeddings", "spark"]
omit = ["tests/*", "*/migrations/*"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
